Revert "lsa: Reduce reclamation latency"
This reverts commit d61002cc33ca2cb5a3c7e8fd2f027cf92bc089d3.

Introduced a regression in row_cache_alloc_stress.

The problem is that reclaim_from_evictable() evicts way too much after
the refactor due to the stop condition not taking into account how
much data was evicted so far and only looking at occupancy of the
minimal segment. This may lead to eviction of the whole region.